requirements.txt
aiofiles==24.1.0
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.7.0
ecdsa==0.18.0
email_validator==2.2.0
fastapi==0.108.0
fastapi-mail==1.4.1
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.4
MarkupSafe==3.0.2
mercadopago==2.2.3
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.1
passlib==1.7.4
pillow==11.0.0
proto-plus==1.25.0
protobuf==4.21.12
pyasn1==0.5.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.2.3
uvicorn==0.25.0
